+++ lib/interfaces/Admin/screens/approved.dart
@@ -124,7 +124,7 @@ Widget getApprovedProducts(Product product,BuildContext context){
                       color: Colors.black54,
                       fontWeight: FontWeight.bold),
                 ),
-                SizedBox(height: 10,),
+                SizedBox(height: 2,),
                 Container(
                   height: 200,
                   width: 200,
@@ -140,7 +140,6 @@ Widget getApprovedProducts(Product product,BuildContext context){
                     ),
                   ),
                 ),
-                SizedBox(height: 5,),
                 Row(
                   children: [
                     ElevatedButton(
@@ -163,12 +162,39 @@ Widget getApprovedProducts(Product product,BuildContext context){
                           fixedSize: const MaterialStatePropertyAll<Size>(Size(80,20)),
                           backgroundColor: MaterialStatePropertyAll<Color>(Colors
                               .red)),
-                      onPressed: () async {
-                        await FirebaseFirestore.instance
-                            .collection('products')
-                            .doc(product.id)
-                            .delete();
-                        print('Product deleted');
+                      onPressed: () {
+                        showDialog(
+                          context: context,
+                          builder: (BuildContext context) {
+                            return AlertDialog(
+                              title: Text("Confirmation"),
+                              content: Text("Are you sure you want to delete this product ?"),
+                              actions: [
+                                TextButton(
+                                  child: Text("Cancel"),
+                                  onPressed: () {
+                                    Navigator.of(context).pop();
+                                  },
+                                ),
+                                TextButton(
+                                  child: Text(
+                                    "Delete",
+                                    style: TextStyle(color: Colors.red),
+                                  ),
+                                  onPressed: () async {
+                                    await FirebaseFirestore.instance
+                                        .collection('products')
+                                        .doc(product.id)
+                                        .delete();
+                                    print('Product deleted');
+                                    //Navigator.of(context).pop();
+                                    Navigator.popAndPushNamed(context, '/approved');
+                                  },
+                                ),
+                              ],
+                            );
+                          },
+                        );
                       },
                       child: Text("Delete"),
                     ),
